Do not check for objective improvement
